Move compose package up
