+++ app/src/main/java/org/fnives/test/showcase/compose/ComposeActivity.kt
@@ -1,4 +1,4 @@
-package org.fnives.test.showcase.ui.compose
+package org.fnives.test.showcase.compose
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import com.google.accompanist.insets.ProvideWindowInsets
-import org.fnives.test.showcase.ui.compose.screen.AppNavigation
+import org.fnives.test.showcase.compose.screen.AppNavigation
 
 class ComposeActivity : AppCompatActivity() {
