+++ app/src/main/java/org/fnives/test/showcase/compose/screen/home/HomeScreenState.kt
@@ -1,4 +1,4 @@
-package org.fnives.test.showcase.ui.compose.screen.home
+package org.fnives.test.showcase.compose.screen.home
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
